Create basic view structure for multiplatform project UI
It is intended to be used for the text-editor which can be used by multiple people at the same time. There are no such view elements already existing so I decided to write my own view element. This view should behave exactly the same on browser and android app so this have to be implemented in the common library. There are currently no ready UI-Frameworks for pure kotlin multiplatform code, so in this commit I experiment with a own framework and my already existing painter-library that supports android, java swing, browser canvas elements.